Reorganize .attribute specs
